fix: merge HubRPCClient and HubClient classes
