+++ app/src/hub.ts
@@ -21,10 +21,10 @@ import MessageModel from '~/flatbuffers/models/messageModel';
 import NameRegistryEventModel from '~/flatbuffers/models/nameRegistryEventModel';
 import { isSignerRemove } from '~/flatbuffers/models/typeguards';
 import { HubInterface, HubSubmitSource } from '~/flatbuffers/models/types';
-import HubRPCClient from '~/hubRpcClient';
 import { Node } from '~/network/p2p/node';
 import { NETWORK_TOPIC_CONTACT, NETWORK_TOPIC_PRIMARY } from '~/network/p2p/protocol';
 import SyncEngine from '~/network/sync/syncEngine';
+import HubRpcClient from '~/rpc/client';
 import Server from '~/rpc/server';
 import BinaryRocksDB from '~/storage/db/rocksdb';
 import Engine from '~/storage/engine';
@@ -266,7 +266,7 @@ export class Hub extends TypedEmitter<HubEvents> implements HubInterface {
     await this.diffSyncIfRequired(message, rpcClient);
   }
 
-  private async diffSyncIfRequired(message: ContactInfoContent, rpcClient: HubRPCClient | undefined) {
+  private async diffSyncIfRequired(message: ContactInfoContent, rpcClient: HubRpcClient | undefined) {
     this.emit('syncStart');
     if (!rpcClient) {
       log.warn(`No RPC client for peer, skipping sync`);
@@ -279,7 +279,7 @@ export class Hub extends TypedEmitter<HubEvents> implements HubInterface {
     return;
   }
 
-  private async getRPCClientForPeer(peer: ContactInfoContent): Promise<HubRPCClient | undefined> {
+  private async getRPCClientForPeer(peer: ContactInfoContent): Promise<HubRpcClient | undefined> {
     /*
      * Find the peer's addrs from our peer list because we cannot use the address
      * in the contact info directly
@@ -308,7 +308,7 @@ export class Hub extends TypedEmitter<HubEvents> implements HubInterface {
     }
 
     if (isIP(addressInfo.value.address)) {
-      return new HubRPCClient(addressInfo.value);
+      return new HubRpcClient(addressInfo.value);
     }
 
     log.info({ peerId: peer.peerIdArray()?.toString() }, 'falling back to addressbook lookup for peer');
@@ -334,7 +334,7 @@ export class Hub extends TypedEmitter<HubEvents> implements HubInterface {
     }
 
     const nodeAddress = addr.multiaddr.nodeAddress();
-    return new HubRPCClient({
+    return new HubRpcClient({
       address: nodeAddress.address,
       family: ipFamilyToString(nodeAddress.family),
       // Use the gossip rpc port instead of the port used by libp2p
